created anime reserve to reduce request frequency
list items are queried by MAL id in our anime database. If not found then it is added to the database and the listing is added to the list in question. If found then information is instead pulled from the anime database so no request needs to be made to API, saving API load.